Apply `SceneWrapper` to the loading scene
